+++ EarthMarket.Shared/DTOs/ProductVariantDto.cs
@@ -10,7 +10,7 @@ namespace EarthMarket.Shared.DTOs
     {
         public Guid Key { get; set; }
         public IDictionary<string,string> ProductVariantAttributeValues { get; set; }
-        public float ProductVariantPrice { get; set; }
+        public double ProductVariantPrice { get; set; }
         public IEnumerable<string> ProductVariantImages { get; set; }
     }
 }
